+++ components/ConfigureSettings/ConfigureSettings.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useEffect } from "react";
 import * as Video from "twilio-video";
 import {
   Button,
@@ -33,9 +33,13 @@ export default function ConfigureSettings({}: ConfigureSettingsProps) {
   const toaster = useToaster();
   const { CONFIGURE_SETTINGS_HEADER, CONFIGURE_SETTINGS_DESCRIPTION } =
     TEXT_COPY;
-  const { localTracks, setLocalTracks, formData, clearTrack } = useVideoStore(
-    (state: VideoAppState) => state
-  );
+  const {
+    localTracks,
+    setLocalTracks,
+    formData,
+    clearTrack,
+    devicePermissions,
+  } = useVideoStore((state: VideoAppState) => state);
   const localVideo = localTracks.video;
   const { identity } = formData;
 
@@ -43,8 +47,8 @@ export default function ConfigureSettings({}: ConfigureSettingsProps) {
   const handleOpen = () => setIsOpen(true);
   const handleClose = () => setIsOpen(false);
   const modalHeadingID = useUID();
-  const { videoInputDevices, audioInputDevices, audioOutputDevices, isMicPermissionGranted, isCameraPermissionGranted } =
-    useDevices();
+  const { videoInputDevices, audioInputDevices, audioOutputDevices } =
+    useDevices(devicePermissions);
 
   function deviceChange(
     deviceID: string,
@@ -81,6 +85,11 @@ export default function ConfigureSettings({}: ConfigureSettingsProps) {
     }
   }
 
+  useEffect(() => {
+    console.log("useEffect > ConfigureSettings");
+    console.log("devicePermissions", devicePermissions);
+  }, [devicePermissions]);
+
   return (
     <Flex hAlignContent={"center"} width="100%">
       <Tooltip text="Configure settings" placement="top">
@@ -105,16 +114,16 @@ export default function ConfigureSettings({}: ConfigureSettingsProps) {
         <ModalBody>
           <Paragraph>{CONFIGURE_SETTINGS_DESCRIPTION}</Paragraph>
           <Stack orientation={"vertical"} spacing="space60">
-            {isCameraPermissionGranted &&
+            {devicePermissions.camera && (
               <VideoPreview
                 identity={identity ?? "Guest"}
                 localVideo={localVideo}
               />
-            }
+            )}
             <Stack orientation="vertical" spacing="space30">
               <Label htmlFor="author">
                 Video{" "}
-                {isCameraPermissionGranted && localVideo === undefined
+                {devicePermissions.camera && localVideo === undefined
                   ? "(disabled)"
                   : localVideo?.isStopped
                   ? "(stopped)"
@@ -123,23 +132,29 @@ export default function ConfigureSettings({}: ConfigureSettingsProps) {
               <Select
                 id="author"
                 onChange={(e) => deviceChange(e.target.value, "video")}
-                defaultValue={isCameraPermissionGranted ? (localVideo?.mediaStreamTrack.id ?? "") : "no-cam-permission"}
+                defaultValue={
+                  devicePermissions.camera
+                    ? localVideo?.mediaStreamTrack.id ?? ""
+                    : "no-cam-permission"
+                }
                 disabled={
-                  videoInputDevices.length < 2 ||
-                  !isCameraPermissionGranted
+                  videoInputDevices.length < 2 || !devicePermissions.camera
                 }
               >
-                {isCameraPermissionGranted ?
+                {devicePermissions.camera ? (
                   videoInputDevices.map((videoInput: MediaDeviceInfo) => (
-                    <Option key={videoInput.deviceId} value={videoInput.deviceId}>
+                    <Option
+                      key={videoInput.deviceId}
+                      value={videoInput.deviceId}
+                    >
                       {videoInput.label}
                     </Option>
-                  )) : (
-                    <Option key="no-cam-permission" value="no-cam-permission">
-                      Camera permissions have not been granted in the browser
-                    </Option>
-                  )
-                }
+                  ))
+                ) : (
+                  <Option key="no-cam-permission" value="no-cam-permission">
+                    Camera permissions have not been granted in the browser
+                  </Option>
+                )}
               </Select>
             </Stack>
             <Stack orientation="vertical" spacing="space30">
@@ -147,23 +162,27 @@ export default function ConfigureSettings({}: ConfigureSettingsProps) {
               <Select
                 id="author"
                 onChange={(e) => deviceChange(e.target.value, "audioInput")}
-                defaultValue={isMicPermissionGranted ? "" : "no-mic-permission"}
+                defaultValue={
+                  devicePermissions.microphone ? "" : "no-mic-permission"
+                }
                 disabled={
-                  audioInputDevices.length < 2 ||
-                  !isMicPermissionGranted
+                  audioInputDevices.length < 2 || !devicePermissions.microphone
                 }
               >
-                {isMicPermissionGranted ? 
+                {devicePermissions.microphone ? (
                   audioInputDevices.map((audioInput: MediaDeviceInfo) => (
-                    <Option key={audioInput.deviceId} value={audioInput.deviceId}>
+                    <Option
+                      key={audioInput.deviceId}
+                      value={audioInput.deviceId}
+                    >
                       {audioInput.label}
                     </Option>
-                  )) : (
-                    <Option key="no-mic-permission" value="no-mic-permission">
-                      Microphone permissions have not been granted in the browser
-                    </Option>
-                  )
-                }
+                  ))
+                ) : (
+                  <Option key="no-mic-permission" value="no-mic-permission">
+                    Microphone permissions have not been granted in the browser
+                  </Option>
+                )}
               </Select>
             </Stack>
             <Stack orientation="vertical" spacing="space30">
@@ -172,14 +191,20 @@ export default function ConfigureSettings({}: ConfigureSettingsProps) {
                 id="author"
                 onChange={(e) => deviceChange(e.target.value, "audioOutput")}
               >
-                {audioOutputDevices.map((audioOutput: MediaDeviceInfo) => (
-                  <Option
-                    key={audioOutput.deviceId}
-                    value={audioOutput.deviceId}
-                  >
-                    {audioOutput.label}
+                {devicePermissions.microphone ? (
+                  audioOutputDevices.map((audioOutput: MediaDeviceInfo) => (
+                    <Option
+                      key={audioOutput.deviceId}
+                      value={audioOutput.deviceId}
+                    >
+                      {audioOutput.label}
+                    </Option>
+                  ))
+                ) : (
+                  <Option key="no-mic-permission" value="no-mic-permission">
+                    Microphone permissions have not been granted in the browser
                   </Option>
-                ))}
+                )}
               </Select>
             </Stack>
           </Stack>
