+++ lib/utils/devices.ts
@@ -1,6 +1,21 @@
+// This function will return 'true' when the specified permission has been denied by the user.
+// If the API doesn't exist, or the query function returns an error, 'false' will be returned.
+export async function getPermissionStatus(name: "camera" | "microphone") {
+  const permissionName = name as PermissionName; // workaround for https://github.com/microsoft/TypeScript/issues/33923
 
-
-
+  if (navigator.permissions) {
+    try {
+      const result = await navigator.permissions.query({
+        name: permissionName,
+      });
+      return result;
+    } catch {
+      return null;
+    }
+  } else {
+    return null;
+  }
+}
 
 export async function getDeviceInfo() {
   const devices = await navigator.mediaDevices.enumerateDevices();
@@ -16,8 +31,12 @@ export async function getDeviceInfo() {
     hasVideoInputDevices: devices.some(
       (device) => device.kind === "videoinput"
     ),
-    isMicPermissionGranted: devices.filter((device) => device.kind === "audioinput").every((d : MediaDeviceInfo) => d.label),
-    isCameraPermissionGranted: devices.filter((device) => device.kind === "videoinput").every((d : MediaDeviceInfo) => d.label)
+    isMicPermissionGranted: devices
+      .filter((device) => device.kind === "audioinput")
+      .every((d: MediaDeviceInfo) => d.label),
+    isCameraPermissionGranted: devices
+      .filter((device) => device.kind === "videoinput")
+      .every((d: MediaDeviceInfo) => d.label),
   };
 }
 
@@ -30,8 +49,6 @@ export const findDeviceByID = (id: string, deviceList: MediaDeviceInfo[]) => {
   }
   return null;
 };
-
-
 
 export const isMobile = (() => {
   if (
